Support setting frequency via serial
To set the frequency, connect via serial and enter e.g.,:
    F1M
followed by enter/return to set the frequency to 1000000Hz (1MHz)

Frequency can be set as a whole number, a whole number of kHz, or a
whole number of MHz.  The commands are case insensitive.  So all
the following set the frequency to 1MHz:
    F1000000
    F1000k
    F1M

No validation is performed and nonsensical things can be requested,
so for example if you enter the nonsensical "F9MMM", you get
"GPS Initialized to output RF at 3800301568"  This specific number
results from overflow of integer arithmetic and is also way out of
spec for the board, since the actual value requested is around 3.8GHz.